+++ TLEdit/tled/undo.h
@@ -26,7 +26,7 @@ void RecordGOMoveStart(GraphicObject* g);
 void RecordGOMoveComplete(GraphicObject* g);
 void RecordChangedProps(GraphicObject* g, PropCellBase* pcp);
 void RecordIrreversibleAct(const char * description);
-void RecordWildfireTCSpread(TSSet&, int old_tcid, int new_tcid);
+void RecordWildfireTCSpread(TSSet&, long old_tcid, long new_tcid);
 void RecordJointCreation(TrackJoint* tj, WPPOINT seg_id);
 void RecordSegmentCut(TrackSeg* ts);
 void RecordSegmentCreation(TrackSeg* ts);
